refactor: DTO 디렉토리 분리
