+++ backend/src/main/java/backend/mulkkam/member/controller/MemberController.java
@@ -1,8 +1,8 @@
 package backend.mulkkam.member.controller;
 
-import backend.mulkkam.member.dto.MemberNicknameModifyRequest;
-import backend.mulkkam.member.dto.MemberNicknameResponse;
-import backend.mulkkam.member.dto.PhysicalAttributesModifyRequest;
+import backend.mulkkam.member.dto.request.MemberNicknameModifyRequest;
+import backend.mulkkam.member.dto.response.MemberNicknameResponse;
+import backend.mulkkam.member.dto.request.PhysicalAttributesModifyRequest;
 import backend.mulkkam.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
