+++ backend/src/main/java/backend/mulkkam/member/dto/request/MemberNicknameModifyRequest.java
@@ -1,4 +1,4 @@
-package backend.mulkkam.member.dto;
+package backend.mulkkam.member.dto.request;
 
 public record MemberNicknameModifyRequest(String memberNickname) {
 }
