+++ backend/src/test/java/backend/mulkkam/member/service/MemberServiceUnitTest.java
@@ -3,8 +3,8 @@ package backend.mulkkam.member.service;
 import backend.mulkkam.member.domain.Member;
 import backend.mulkkam.member.domain.vo.Gender;
 import backend.mulkkam.member.domain.vo.MemberNickname;
-import backend.mulkkam.member.dto.MemberNicknameModifyRequest;
-import backend.mulkkam.member.dto.PhysicalAttributesModifyRequest;
+import backend.mulkkam.member.dto.request.MemberNicknameModifyRequest;
+import backend.mulkkam.member.dto.request.PhysicalAttributesModifyRequest;
 import backend.mulkkam.member.repository.MemberRepository;
 import backend.mulkkam.support.MemberFixtureBuilder;
 import java.util.Optional;
